Refactor to use an abstract class
