+++ src/main/java/com/loopingz/BasicSmtpRelay.java
@@ -1,26 +1,21 @@
 package com.loopingz;
 
 import com.amazonaws.util.StringUtils;
-import org.subethamail.smtp.helper.SimpleMessageListener;
-import org.subethamail.smtp.server.SMTPServer;
 
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
 import java.util.Properties;
 
-public class BasicSmtpRelay implements SimpleMessageListener {
+public class BasicSmtpRelay extends SmtpRelay {
 
-    private DeliveryDetails deliveryDetails;
     private Properties props;
     private boolean authRequest;
 
     BasicSmtpRelay(DeliveryDetails deliveryDetails) {
-        this.deliveryDetails = deliveryDetails;
+    	super(deliveryDetails);
 
         props = new Properties();
         props.put("mail.smtp.starttls.enable", "true");
@@ -34,11 +29,6 @@ public class BasicSmtpRelay implements SimpleMessageListener {
             props.put("mail.smtp.auth", "false");
             authRequest = false;
         }
-    }
-
-    @Override
-    public boolean accept(String from, String to) {
-        return true;
     }
 
     @Override
@@ -68,14 +58,5 @@ public class BasicSmtpRelay implements SimpleMessageListener {
                     });
         }
         return Session.getDefaultInstance(props);
-    }
-
-    void run() throws UnknownHostException {
-        SMTPServer.Builder builder = new SMTPServer.Builder();
-        builder.bindAddress(InetAddress.getByName(deliveryDetails.getBindAddress()))
-                .port(deliveryDetails.getPort())
-                .simpleMessageListener(this);
-        SMTPServer smtpServer = builder.build();
-        smtpServer.start();
     }
 }
